Move BP+ folder
