+++ coins/monero/src/ringct/bulletproofs/plus/scalar_vector.rs
